backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.35.3
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.0
slowapi==0.1.9
elevenlabs==1.5.0
